Create Fibonacci_goldenratio.py
Time complexity : O(1)O(1). Constant time complexity since we are using no loops or recursion and the time is based on the result of performing the calculation using Binet's formula.

Space complexity : O(1)O(1). The space used is the space needed to create the variable to store the golden ratio formula.